Add files via upload
I develop an e-commerce website to sell various kinds of mobile and smartphones. I have developed an e-commerce mobile store named LOKI following the guidelines provided. My website contains many pages but mainly Home, Products, Blog, Research, and Contact us are the main attraction of my website. My website sells various kinds of smartphones ranging from mid-range to flagship. Various brands such as apple, Samsung, Vivo, Asus, and Xiaomi are available in my store.  
Each web page of my site contains a navigation bar at the top and a footer bar at the bottom of the page to navigate throughout the website with ease. The main page of my website is the Home page. It consists of an image slider just below the navigation bar at the top which I developed using HTML and CSS. Then after that, it has a section that appeals to the customers to buy a specific phone with the help of attractive taglines. Then there is another section named your trusted brands which shows the customers that we have all the important and most selling brands smartphones on our website. After that, there is a section called our new additions which shows the latest phones in the market are also available in our store and this section consists of description section which shows the specification of the fresh phones. Then there is a section named as Benefits of joining us which displays the services which we provide and the special offers like free delivery which can attract the customers to take advantage of our offers. At last, there is a footer section that consists of the link to all the pages of our site and in the middle, there is a link that directs the customers to the privacy policies and terms and conditions of our store and website. 
The second page of my website is the products page. This page includes an image banner just below the navigation bar. Then there is a section named Hot Picks which displays the current most selling products on our website which can motivate a customer to buy those products. There is the section for the apple brand which shows all the apple products available in the store and then similarly there are sections for the brand Xiaomi, Samsung, Asus, and Vivo. The third page of the website is the Blog page. On this page, I have written a blog on how technology has affected our daily life positively and negatively. I have written this blog including various relatable images. The fourth page of my website is the Research page. On this page, I have shown the websites from which I have taken references to develop my website. The last page of my website is to contact us from where they can contact us in case of any faulty delivery or visit our physical store and can even give feedback to us on our services.
I have developed my website using HTML, CSS, and JavaScript. I have used HTML to build the fundamental building block of my web pages and to control the layout of my contents. HTML helped me to provide the structure for my web pages. On other hand, I applied CSS to style my webpages elements. I targeted various screen sizes to make my web pages responsive as well. CSS helped me to provide that look and feel to my website. I used JavaScript to make my website interactive with the customers. I added different functions and features to my website using JavaScript. 
In my opinion, I have made a very decent and well-managed website. The main point for me is that I got an opportunity to learn more about web development through this coursework and it has boosted my confidence that this will help to make the future better. It helped me to enhance my skills more and taught me to develop an interactive website. Developing an interactive website is not easy work and the coursework like this is helping me to learn more.